+++ backend/src/main/java/cn/etarch/mao/notification/task/mapper/TaskNotificationDeliveryMapper.java
@@ -1,7 +1,0 @@
-package cn.etarch.mao.notification.task.mapper;
-
-import cn.etarch.mao.notification.task.entity.TaskNotificationDelivery;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-
-public interface TaskNotificationDeliveryMapper extends BaseMapper<TaskNotificationDelivery> {
-}