+++ backend/src/main/java/cn/etarch/mao/permission/mapper/RolePermissionMapper.java
@@ -1,9 +1,0 @@
-package cn.etarch.mao.permission.mapper;
-
-import cn.etarch.mao.permission.entity.RolePermission;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface RolePermissionMapper extends BaseMapper<RolePermission> {
-}