+++ backend/src/main/java/cn/etarch/mao/permission/mapper/UserRoleMapper.java
@@ -1,9 +1,0 @@
-package cn.etarch.mao.permission.mapper;
-
-import cn.etarch.mao.permission.entity.UserRole;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface UserRoleMapper extends BaseMapper<UserRole> {
-}