+++ backend/src/main/java/cn/etarch/mao/preference/mapper/UserWeixinPreferenceMapper.java
@@ -1,9 +1,0 @@
-package cn.etarch.mao.preference.mapper;
-
-import cn.etarch.mao.preference.entity.UserWeixinPreference;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface UserWeixinPreferenceMapper extends BaseMapper<UserWeixinPreference> {
-}