+++ backend/src/main/java/cn/etarch/mao/session/activity/SessionActivityMapper.java
@@ -1,8 +1,0 @@
-package cn.etarch.mao.session.activity;
-
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface SessionActivityMapper extends BaseMapper<SessionActivity> {
-}