+++ backend/src/main/java/cn/etarch/mao/user/mapper/GitCredentialMapper.java
@@ -1,9 +1,0 @@
-package cn.etarch.mao.user.mapper;
-
-import cn.etarch.mao.user.entity.GitCredential;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface GitCredentialMapper extends BaseMapper<GitCredential> {
-}